Add username argument to vcloud-login
Allows a username to be passed to `vcloud-login` on the command line. Rather
than always prompting for it.

This makes it easier to use against Platform1, where the username is a UID
and has to be suffixed with the organisation UID, which is annoying when the
auth token expires every hour or so.